+++ src/app/providers.tsx
@@ -6,9 +6,15 @@ import { ConfettiProvider } from "@/contexts/confetti-context";
 import { PrivyProvider } from "@privy-io/react-auth";
 import { WagmiProvider, createConfig } from "@privy-io/wagmi";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { NextIntlClientProvider } from "next-intl";
+import { usePathname, useSearchParams } from "next/navigation";
+import { usePostHog } from "posthog-js/react";
+import { Suspense, useEffect } from "react";
 import { arbitrumSepolia, aurora, base, matchain } from "viem/chains";
 import { http } from "wagmi";
-import { NextIntlClientProvider } from 'next-intl';
+
+import posthog from "posthog-js";
+import { PostHogProvider as PHProvider } from "posthog-js/react";
 
 const chainConfig = createConfig({
   chains: [arbitrumSepolia, aurora, turboChain, base, matchain],
@@ -22,11 +28,11 @@ const chainConfig = createConfig({
 });
 const queryClient = new QueryClient();
 
-export default function Providers({ 
-  children, 
+export default function Providers({
+  children,
   messages,
-  locale 
-}: { 
+  locale,
+}: {
   children: React.ReactNode;
   messages: any;
   locale: string;
@@ -52,10 +58,60 @@ export default function Providers({
       >
         <QueryClientProvider client={queryClient}>
           <WagmiProvider config={chainConfig}>
-            <ConfettiProvider>{children}</ConfettiProvider>
+            <PostHogProvider>
+              <ConfettiProvider>{children}</ConfettiProvider>
+            </PostHogProvider>
           </WagmiProvider>
         </QueryClientProvider>
       </PrivyProvider>
     </NextIntlClientProvider>
+  );
+}
+
+export function PostHogProvider({ children }: { children: React.ReactNode }) {
+  useEffect(() => {
+    posthog.init(process.env.NEXT_PUBLIC_POSTHOG_KEY as string, {
+      api_host: process.env.NEXT_PUBLIC_POSTHOG_HOST || "https://eu.i.posthog.com",
+      person_profiles: "identified_only", // or 'always' to create profiles for anonymous users as well
+      capture_pageview: false, // Disable automatic pageview capture, as we capture manually
+    });
+  }, []);
+
+  return (
+    <PHProvider client={posthog}>
+      <SuspendedPostHogPageView />
+      {children}
+    </PHProvider>
+  );
+}
+
+function PostHogPageView() {
+  const pathname = usePathname();
+  const searchParams = useSearchParams();
+  const posthog = usePostHog();
+
+  // Track pageviews
+  useEffect(() => {
+    if (pathname && posthog) {
+      let url = window.origin + pathname;
+      if (searchParams.toString()) {
+        url = `${url}?${searchParams.toString()}`;
+      }
+
+      posthog.capture("$pageview", { $current_url: url });
+    }
+  }, [pathname, searchParams, posthog]);
+
+  return null;
+}
+
+// Wrap PostHogPageView in Suspense to avoid the useSearchParams usage above
+// from de-opting the whole app into client-side rendering
+// See: https://nextjs.org/docs/messages/deopted-into-client-rendering
+function SuspendedPostHogPageView() {
+  return (
+    <Suspense fallback={null}>
+      <PostHogPageView />
+    </Suspense>
   );
 }